net: phylink: pretty-print Kconfig name and status of PHY missing driver
Figuring out, based on the PHY ID, which driver is appropriate is not
always easy. This involves walking the source code and masking the PHY
ID we can get from sysfs or other places with the phy_driver phy_id_mask.

Typically users care about the PHY ID when there isn't a matching driver
already loaded. That's a problem, because exactly then is when the
kernel also has no idea what PHY driver might be adequate for a PHY ID.

My idea is to move each driver's struct mdio_device_id to a separate
stub file, always compiled into phylib. I've implemented this for the
Marvell and BCM84881 drivers, but it will have to scale for all PHY
drivers eventually. The good part is that it can be rolled out gradually,
and not all PHY drivers need to use this infrastructure from the get go.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>