net: dsa: replay tag_8021q VLANs in multi-chip trees on dsa_tag_8021q_register()
tag_8021q does not work out of the box on trees with multiple switches.

The notifier system does not achieve its purpose (every switch programs
itself on the DSA and CPU ports with the VLANs set on the notified
switch port) at probe time.

This is because the switches call dsa_tag_8021q_register() during
their ds->ops-> setup() method or later, and this is done sequentially
(dsa_register_switch() runs under dsa2_mutex). I.e. when the first
switch in the tree probes, DSA_NOTIFIER_TAG_8021Q_VLAN_ADD is correctly
emitted, but the other switch does not listen for it yet, because it
hasn't called dsa_tag_8021q_register() yet.

Calling dsa_tag_8021q_register() before dsa_register_switch() is hard,
because the notifier system just isn't ready.

Introduce one more notifier to replay the VLAN context of the other
switches from the tree. In the example above, this is useful for the
second switch to find out what it missed out on.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>