net: dsa: sja1105: prepare sja1105_ptp_txtstamp_skb() to timestamp packets without a clone
PTP over IP needs to be trapped by SJA1105 to the CPU, and this means
configuring FDB entries with TAKETS=1 for them, regardless of whether
user space wants timestamps (for event messages) or not (for general
messages).

These FDB entries need to have their DESTPORTS field updated per packet
transmitted from the CPU, and this means ensuring that no PTP packet is
in flight concurrently with that update.

Since FDB entries are not management routes as set up by
sja1105_mgmt_route_install(), we have no sja1105_mgmt_route_poll() to
make sure that the PTP packets have been transmitted.

The only thing we can use as a transmit confirmation is
sja1105_ptp_txtstamp_skb(), but we have to prepare it for the case where
we don't actually care for the timestamp, we just need to wait.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>