net: dsa: sja1105: create internal versions of sja1105pqrs_fdb_add()/_del()
FDB entries in the P/Q/R/S switch generation have extra fields: TAKETS
and TSREG, which allows hardware timestamping for custom destination MAC
addresses, in both receive and transmit directions.

Received packets which match an FDB entry with TAKETS=1 are trapped to
the CPU. Transmitted packets which match an FDB entry with TAKETS=1 are
forwarded to the ports in the DESTPORTS field of that FDB entry, and
a timestamp is collected in the TSREG register for each port that the
packet is sent to.

This FDB entry needs to be reprogrammed per packet, because we don't
want the stack to think it's sending a Sync packet to one port of a
boundary clock and in reality that packet to egress multiple ports.

So initially, the FDB entries should be added towards no egress port
(DESTPORTS=0), which is sufficient for reception. The transmission path
will find the FDB entry and update the DESTPORTS field.

So we need to customize this internal API to allow the creation of
FDB entries towards no port, so replace "int port" with "unsigned long
destports".

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>