net: dsa: tag_sja1105: prepare data path for extended L2 timestamps
Extended L2 addresses are FDB entries with TAKETS=1, which the switch
driver configures independently of the tagger.

They are relevant because these packets are trapped to CPU in the
receive path, so we shouldn't set skb->offload_fwd_mark, and the
software bridge should decide if they must be forwarded in software or
not.

Also, on TX, a special procedure is necessary, because the DESTPORTS
field of that FDB entry needs to be updated over SPI with the index of
the user port that the packet should be routed to.

Communication between the tagging protocol driver and the switch driver
cannot happen through exported symbols. At the same time, tag_sja1105
should be able to be paired to other switch drivers as well, like
dsa_loop.

To support this mechanism when connected to the sja1105 switch driver,
introduce a function pointer, populated by the switch driver at tagging
protocol connect time, which is used as a query mechanism by the tagging
protocol for TAKETS=1 FDB entries on both receive and transmit.

The result of this lookup is cached, both on RX and TX, in
SJA1105_SKB_CB :: needs_extended_l2_tstamp. The switch driver is free to
use it during the deferred xmit procedure, but since the FDB entries can
come and go concurrently with the packet data path, the model is only
that of "eventual consistency". The caching in skb->cb is only so that
consistent decisions are taken throughout the skb lifetime, and the
presence of the FDB entry is only evaluated once.

When ports are under a bridge, the source port and switch id for
timestampable packets are initially imprecisely determined, only to be
fully known when the meta frame arrives.

This is problematic on multiple levels.

First, we need to assume that the driver either sets up hardware
timestamping for all switches in the fabric, or for none.
sja1105_any_extended_l2_trap() tests for this. Otherwise we do not know
whether to expect a meta frame for the packet or not.

Second, with cross-chip bridging, the stampable packet can initially be
assumed to belong to a different switch than that from which the meta
frame will later arrive. We need to relax the existing condition that
the two have the same provenance.

The switching fabric guarantees it does not reorder timestampable frames
with their meta frames even across multiple hops, because they are sent
in order, and the QoS class of the timestampable frames is >= the QoS
class of the meta frames.

There exists a risk of incorrect matching of a timestampable frame to a
different meta frame than the one that the hardware produced for it.
This occurs under packet loss, specifically if the hardware produces
packets A, A', B, B', and A' and B are lost, B' could be associated with
A.

This problem also exists for regular L2 timestamping, but is less worse
there because we know the source port/switch ID of packet A, and we can
compare it against B'. Here, only the time-based comparison algorithm
can protect against that. If A was received more than 5 ms ago, it
cannot be matched with B' and is discarded instead.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>