net: dsa: connect/disconnect from PHY at runtime
There are use cases where certain PHYs of a switch may be part of EX
zones, and the system is required to be able to go on without them.
In this situation, the switch itself is not part of the EX area, and is
assumed to be part of the system that maintains power at all times.

It is known by a few seconds ahead of time that the PHY will lose power,
so a usage model which is sufficient for that use case is to bring the
port down and unbind the driver from the PHY device. Then, any attempt
to bring the port up will fail due to the PHY not responding over MDIO.
When the PHY will start responding again, the port can be brought up
again. This behavior is not currently possible because DSA connects to
the PHY at probe time, but other drivers connect at ndo_open() time,
offering compatibility with that kind of behavior already.

I am aware that switching from one operating mode to another creates
interesting backwards compatibility concerns, like for example ethtool
-s commands before the port is up will not be talking to the PHY, when
before this patch they were.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>