net: dsa: delete dead code in dsa_port_touch()
The logic to find an already existing struct dsa_port in the tree's
dst->ports will never find it, after commit 6dc43cd3aae0 ("net: dsa: Fix
use-after-free in probing of DSA switch tree"). This is because the
original design - technically introduced in commit 05f294a85235 ("net:
dsa: allocate ports on touch") but practically somewhere between the
lines - relied on use-after-free.

If a switch succeeds probing, it leaves its ports in dst->ports. If it
fails to probe, it must release its ports from dst->ports. That is what
commit 6dc43cd3aae0 ("net: dsa: Fix use-after-free in probing of DSA
switch tree") does. Therefore, since dsa_switch_probe() only runs until
successfully probed and never afterwards, it means that dsa_port_touch()
will never find a pre-existing struct dsa_port in dst->ports.

This matters because I'd like to change the operation ordering at probe
time. Currently we have:

dsa_switch_probe()
-> dsa_switch_get_tree() // 1
-> dsa_switch_touch_ports() // 2
-> dsa_switch_parse_ports_of() // 3
-> dsa_tree_setup() // 4

I'd like to move 3 before 1, but 3 depends on port creation done by 2.
In itself, 2 not only creates ports, but also adds them to dst->ports
created by 1 - so it depends on that. So I'd like to split
dsa_switch_touch_ports() into a pure "creation" phase which can go
before 1, and a "touching" phase which adds them to dst->ports.

That splitting cannot go forward because of this logic, which assumes
ports shouldn't be created if they are already part of dst->ports.
But as explained, it is only a theoretical possibility, and we should
remove the logic, to permit the rest of the refactoring to take place.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>