net: dsa: split port creation out of dsa_switch_touch_ports()
The present DSA switch probing path is as follows:

dsa_switch_probe()
-> dsa_switch_get_tree() // 1
-> dsa_switch_touch_ports() // 2
-> dsa_switch_parse_ports_of() // 3
-> dsa_tree_setup() // 4

I'd like to move 3 before 1, but 3 depends on port creation done by 2.
In itself, 2 not only creates ports, but also adds them to dst->ports
created by 1 - so it depends on that.

To break up that dependency, we must recognize that it isn't circular,
it's just that the pieces of 2 that 3 depends on (creating the ports)
are not the same that depend on 1 (adding the ports to dst->ports, or
"touching" them - strange name).

So I'd like to break up 2 into a simpler dsa_switch_create_ports() that
can go before 1, and the remainder which will continue having the name
dsa_switch_touch_ports(), and its present location, for now.

It is presently impossible to iterate over a switch's ports if they
haven't been added to dst->ports, since they are discrete,
non-contiguous chunks of memory. The current way implies iterating
through dst->ports and filtering only those ports with dp->ds == ds.
But because we need to enforce a split, we need another way to iterate
through the switch's ports.

Here we propose a contiguous allocation of all switch ports, so that we
can iterate through them as an array. The tree code continues to have
pointers to individual ports as if that weren't the case. We just need
an extra pointer, per switch, to the first port, and we already have the
port count as ds->num_ports.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>