net: dsa: use tree components to determine whether it is complete
Currently, the one and only indication that the routing table is
incomplete (i.e. there are switches remaining to probe) is that
following at least one "link" phandle using dsa_tree_find_port_by_node()
results in a struct dsa_port not found in dst->ports.

This is highly inconvenient if we want to remove the dependency upon
the "link" phandle, and just specify a single "cascade" to the adjacent
port.

Rework the "complete" indication to be based on the components
calculated from BFS on the device tree, which works in both cases.

This makes dsa_port_setup_routing_table() return hard errors, including
-ENOMEM which was previously treated as "incomplete tree".

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>