net: dsa: resolve tagging protocol at dsa_tree_bind_switch() time
The idea that dsa_switch_probe() should just minimally parse the switch
and immediately let the tree do the rest by calling dsa_switch_get_tree()
hits an impediment in dsa_port_parse_cpu(), because there is a strong
dependency on the switch tree which has not yet been created. We want to
impose all sorts of restrictions about all switches in the same tree
using the same tagging protocol (when that is defined per CPU port).

Just minimally parse CPU ports like we parse DSA ports, and delay
resolving the tagging protocol until the switch binds to the tree.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>