net: dsa: run BFS to create a list of switch tree components
The wider goal is to have an UAPI-visible kernel object whose lifetime
is as long as at least one DSA switch is probed. The tree object builds
its own view of the rest of the tree by performing BFS starting from the
OF node of the initial switch that triggered its creation.

The component API was considered, where the tree virtual device would
implement struct component_master_ops, and individual switch devices
would implement struct component_ops. But this does not seem entirely
adequate for DSA's long term goals, as discussed here:
https://lore.kernel.org/netdev/20240918111008.uzvzkcjg7wfj5foa@skbuf/

This change merely explores the device tree, constructing the list of
components, one for the OF node of each switch.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>