net: dsa: bind switches to the tree
Expand struct dsa_tree_component with a state (currently either
"unbound" or "bound"), and a pointer to the struct dsa_switch associated
with that component's OF node, if bound.

Only allow switches to bind to the tree if the tree has already
identified them as components, based on the BFS algorithm.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>