net: dsa: sja1105: move H topology loop prevention mechanism to core
sja1105 has logic which is a bit too complex for a simple driver to
have, and it is the only driver to walk over the tree's routing table
without any helper function.

The routing table's data structures are in the process of changing, so
just abstract the logic into a new function (with a very long, albeit
hopefully descriptive name).

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>