net: dsa: sja1105: don't circumvent the exclusivity of the reset GPIO
Commit 33e1501f5a5f ("net: dsa: sja1105: don't keep a persistent
reference to the reset GPIO") permitted a functionality I was completely
unaware of. Unless GPIOD_FLAGS_BIT_NONEXCLUSIVE is specified,
gpiod_get_optional() by default gets us exclusive access to a GPIO
descriptor. Though, by calling gpiod_put() as soon as we assert the
reset line, we circumvent that and allow another consumer to acquire
it and do the same.

Even more strangely, shared reset GPIOs do happen to work with the
sja1105 driver as currently written, as long as they're all shared just
among sja1105 devices. This is because of the barrier from the DSA
framework, which waits for the routing table to be complete before
calling ds->ops->setup() for all switches. Since we do all configuration
from ds->ops->setup() and none from the bare sja1105_probe() method
(except for asserting the hardware reset line and reading the device
ID), technically switch A, by asserting the shared reset line at a time
when switch B has finished probing, does not negatively impact it, since
switch B's driver hasn't committed any configuration changes yet.

This is all about to change with a new probing scheme in DSA, where
switches are allowed to have life times independent of one another,
and the tree reconfigures itself and the routing table as switches
come and go. In that usage pattern, the probing of a new switch is
supposed to not affect the operation of a pre-existing one.

Supporting the shared reset GPIO use case was unintentional, and I
would like to take that back, since it creates unnecessary pressure on
the implementation, for a benefit that isn't clear. I don't consider the
use case sane, and I would recommend to whomever has such a hardware
design to change it. No other DSA drivers support this use case: they
all use devm_gpiod_get_optional() - for a reason which initially wasn't
clear to me, but now is.

The failure introduced here looks like this:

[    2.145381] sja1105 spi1.0: error -EBUSY: Failed to exclusively acquire reset GPIO
[    2.153033] sja1105: probe of spi1.0 failed with error -16

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>