net: dsa: detect duplicate component ports and keep them sorted
Since the device tree might keep component ports in an unsorted order
but we print them to /sys/class/dsa_switch_tree/.../components, it
becomes necessary to sort the list. Thanks to the previous change we
have the component port index, so we can do it. Also detect duplicates
and fail the probe.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>