net: dsa: tag_sja1105: prepare receive path for meta frames from downstream switches
The hardware sends meta frames as VLAN-untagged. But when there is an
upstream switch that receives these and forwards them on towards the
CPU, it encapsulates them in SJA1105_DEFAULT_VLAN (the PVID on the
downstream cascade port).

Prepare the software receive path for that case, by advancing the
"buffer" pointer in sja1105_meta_unpack() to account for the possible
presence of a VLAN header.

Note that the NETIF_F_HW_VLAN_CTAG_RX feature in the conduit driver
might mask away the need for this, when the port is under a VLAN-aware
bridge, because the VLAN will have TPID ETH_P_8021Q, which is going to
be placed in the hwaccel area, and the existing code will work as is.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>