net: dsa: introduce dsa_tree_for_each_switch() iterator
Currently one cannot iterate uniquely through the dsa_switch elements of
a dsa_switch_tree, one has to either emit a cross-chip notifier or walk
through dst->ports using "dp" elements and filter unique dp->ds values.

With the introduction of the dst->components list, we can make that
possible, since a component is the device tree "fossil" of a switch, and
it can either point directly to a dsa_switch (if the component is bound)
or can be skipped otherwise.

I could not figure out how to hide the struct dsa_tree_component
required for walking dst->components inside the macro, so currently it
is the responsibility for the caller to also declare that.

In fact, the component type can be hidden, but then, the other line in
the initializer section for the "for" loop, i.e. "_ds = _component->ds",
is interpreted as if "_ds" is another variable declaration of the type
struct dsa_tree_component *, which it is not.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>