net: dsa: tag_sja1105: reimplement meta state machine using tree-wide skb stack
The SJA1105P/Q/R/S TCAM-based timestamping mechanism traps timestampable
frames to the CPU without making any note regarding which source port or
switch ID they came from. That is later known when the meta frame
carrying the partial RX timestamp for that timestampable frame arrives,
but the problem is circular, because one cannot say with certainty that
the meta frame is intended for a certain timestampable frame, because we
don't know the source port and switch ID of the timestampable frame.

This already makes the existing tagger_private->stampable_skb mechanism
inadequate, because when we receive a meta frame, it can be located on a
different switch than the original switch which we thought had received
the timestampable frame (we initially perform imprecise RX for that).

Additionally, the guarantees about consecutiveness of timestampable
frames and their meta follow-ups don't extend to downstream switches in
a daisy chain topology. They can't: the top-level switch in that
scenario is simply transporting those timestampable and meta frames as
Ethernet payload. If it receives a PTP frame on another user port at the
"wrong" time, it simply has to break up the frames it transports from
the downstream cascade port, simply in order to enforce its own
guarantees about local consecutiveness.

This means we can end up receiving, at the tagger layer, two stampable
skbs (or more) without having received any meta frame. When the single
tagger_private->stampable_skb pointer was per switch, that was fine and
supported. But now it's not, due to the imprecise reception of PTP over
UDP. So we need to loosen that implementation to now become a stack
(LIFO) of timestampable frames which were received across the entire
tree (because we don't initially know where they came from).

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>