net: dsa: reimplement CPU port assignments based on components
The component layer needs to know the mapping between DSA ports and CPU
ports in order for the upstream/downstream relationship to be
determined. A DSA port is upstream if it is the switch's routing port
towards its dedicated CPU port. We need to have that knowledge based on
just component information from the routing table, so that we can use
the routing table in intermediary states where the tree has not fully
probed.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>