net: dsa: lantiq_gswip: write the index of the Active VLAN entry for VLAN-unaware
gswip_vlan_add_aware() and gswip_vlan_add_unaware() differ in few
aspects, one of them is what they write to GSWIP_PCE_DEFPVID(port).

This seems to be an index into the Active VLAN table, since in both
cases they assign idx = gswip_vlan_active_create() if the VLAN table
entry wasn't previously found.

However, the VLAN-unaware variant assumes that this index will be zero,
which does not seem correct. The gswip_vlan_active_create() function
will return an idx starting with at least priv->hw_info->max_ports,
whereas 0 seems to be an unknown index. The submitter had this to say
about it:
| This table index 0 is behaving somehow special I just excluded it from
| normal operations for now.
https://patchwork.ozlabs.org/project/netdev/patch/20190501204506.21579-3-hauke@hauke-m.de/

but it doesn't seem to be all that excluded in this context.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>