net: dsa: xrs700x: reject unsupported HSR configurations
The fact, discussed to death here, but nothing came out of it:
https://lore.kernel.org/netdev/20240620090210.drop6jwh7e5qw556@skbuf/

is that the xrs700x.c driver only supports offloading HSR_PT_SLAVE_A and
HSR_PT_SLAVE_B (which were the only port types at the time the offload
was written, _for this driver_).

Up until now, the API did not explicitly tell offloading drivers what
port has what role. So xrs700x can get confused and think that it can
support a configuration which it actually can't (see the table in the
link attached).

Now that we have explicit hsr_get_port_type() API, let's use that to
work around the unintended semantical changes of the offloading API
brought by the introduction of interlink ports in HSR.

Fixes: 5055cccfc2d1 ("net: hsr: Provide RedBox support (HSR-SAN)")
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>