gpio: mpsse: add MPSSE protocol acceleration API
Some pins of the FTDI GPIO controller USB device have the ability to be
more than just GPIOs. This function is the so-called Multi-Protocol
Synchronous Serial Engine (MPSSE), described in FTDI public application
note AN_135 (FTDI MPSSE Basics).

Essentially, channels A and B of some devices (FT2232H, FT4232H) have
the ability to map GPIO pins:
- 0 as a "data out" pin
- 1 as a "data in" pin
- 2 as a "clock" pin
- 3 as a "chip select" pin

This is further usable to accelerate some serial protocols (SPI, I2C,
JTAG, MDIO etc) in a way that is much more efficient than bit banging.
For example, instead of toggling the "data out" and "clock" pins as
GPIO, using the MPSSE commands we just give the device an octet, and it
will toggle the clock pin by itself until it is transmitted.

The existence of such MPSSE accelerated functions doesn't change much in
the FTDI GPIO controller driver. All protocols (SPI, I2C etc) are
software-defined, and as such will be separate platform drivers, defined
in the root (/) OF node (because otherwise we have no way to know how a
particular board intends the MPSSE pins to be used).

A protocol driver gets access to the MPSSE engine via an OF phandle, and
the GPIO pins via standard 'gpios' phandles. It uses the GPIO API just
to set pin directions and logic levels (active high/low). Then we need
to export a small set of API functions for protocol operations:
- mpsse_write_data()
- mpsse_read_data()
- mpsse_set_clock()
- mpsse_flush()

so that the protocol driver can implement its specific protocol.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>