usb: introduce find_usb_interface_by_of_node()
Allow other kernel drivers to find USB interfaces using phandles to the
OF nodes that the USB bus already creates for individual devices and
interfaces.

usb_bus_type is not exported outside of drivers/usb/core/usb.h, so we
need to create a wrapper function around bus_find_device_by_of_node() in
a place that's also in the USB core, and export that symbol.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>