spi: mpsse: new driver
This is a protocol driver for FTDI USB chips which have a MPSSE
interface (generic accelerator for synchronous protocols, see
https://www.ftdichip.com/Documents/AppNotes/AN_135_MPSSE_Basics.pdf).

The decision is to make the SPI protocol driver OF-based, because:
- The MPSSE is a generic serial protocol accelerator, and we don't know
  ahead of time which protocol should be spoken on the FTDI MPSSE pins.
  Other than SPI, it could as well be I2C, JTAG, MDIO, etc etc.
  Furthermore, there may be more than 1 MPSSE channel in one USB chip.
  Describing the protocol in the device tree seems to be adequate - it
  is describing the electrical connections made on that USB chip in the
  truest sense.
- The SPI protocol driver shouldn't own the entire USB interface. There
  are 16 GPIO pins on an interface, and the non-accelerated MPSSE pins
  can still be used as GPIOs for whatever other purpose.
- We may need to probe very complex SPI children, which are heavily OF
  dependent, like DSA switches.

There are downsides to this architectural choice, like for example that
device trees are somewhat too static for USB's nature (the FTDI chips
are not on-board, but many times are part of the board we want to
control, like a development board with a DSA switch on it). Another one
is that although this SPI protocol driver nicely returns -EPROBE_DEFER
until the USB device is plugged in, things are not so smooth when it
gets unplugged. Normally, we'd want to unbind and rebind the SPI
protocol driver in order for the state machine to rerun from the
beginning, but that is currently not handled.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>