arm64: dts: add NXP SJA1110-MGS-EVM board
https://www.nxp.com/design/design-center/development-boards-and-designs/SJA1110-MGS-EVM

This is connected to the LS1028A-RDB in what is somewhat of an arbitrary
choice - as this is the host board I connected it to, for testing.
However, there are macros through which it can be essentially connected
to anything with an USB port and an Ethernet port, as long as those have
device tree descriptions.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>