net: ethtool: don't reference ethnl_header_policy_phy from ethnl_parse_header_dev_get()
We only use ethnl_header_policy_phy in conjunction with its ARRAY_SIZE(),
that is, the contents does not matter.

The code is correct, however, ethnl_header_policy_phy is a policy that
validates which attributes are permitted for the specific command, and
in the generic ethnl_parse_header_dev_get() that doesn't matter.

Replace the ARRAY_SIZE() with expressions that depend upon
ETHTOOL_A_HEADER_MAX instead. This should avoid some churn when more and
more policies for the header are introduced.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>